refactor
interfaces moved to package `core`
