+++ app/src/androidTest/java/com/forrestguice/suntimeswidget/calculator/SuntimesCalculatorTest.java
@@ -23,6 +23,9 @@ import android.support.test.InstrumentationRegistry;
 import android.support.test.runner.AndroidJUnit4;
 import android.test.RenamingDelegatingContext;
 
+import com.forrestguice.suntimeswidget.calculator.core.Location;
+import com.forrestguice.suntimeswidget.calculator.core.SuntimesCalculator;
+
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
