+++ app/src/main/java/com/forrestguice/suntimeswidget/LocationConfigView.java
@@ -106,7 +106,7 @@ public class LocationConfigView extends LinearLayout
 
     public boolean isInitialized() { return isInitialized; }
 
-    public com.forrestguice.suntimeswidget.calculator.Location getLocation()
+    public com.forrestguice.suntimeswidget.calculator.core.Location getLocation()
     {
         String name = text_locationName.getText().toString();
         String latitude = text_locationLat.getText().toString();
@@ -138,7 +138,7 @@ public class LocationConfigView extends LinearLayout
             units = WidgetSettings.PREF_DEF_GENERAL_UNITS_LENGTH;
         }
 
-        return new com.forrestguice.suntimeswidget.calculator.Location(name, latitude, longitude, altitude, units == WidgetSettings.LengthUnit.METRIC);
+        return new com.forrestguice.suntimeswidget.calculator.core.Location(name, latitude, longitude, altitude, units == WidgetSettings.LengthUnit.METRIC);
     }
 
     public WidgetSettings.LocationMode getLocationMode()
@@ -310,7 +310,7 @@ public class LocationConfigView extends LinearLayout
         @Override
         public void updateUI(Location... locations)
         {
-            DecimalFormat formatter = com.forrestguice.suntimeswidget.calculator.Location.decimalDegreesFormatter();
+            DecimalFormat formatter = com.forrestguice.suntimeswidget.calculator.core.Location.decimalDegreesFormatter();
             text_locationLat.setText( formatter.format(locations[0].getLatitude()) );
             text_locationLon.setText( formatter.format(locations[0].getLongitude()) );
             text_locationAlt.setText( formatter.format(locations[0].getAltitude()) );
@@ -353,7 +353,7 @@ public class LocationConfigView extends LinearLayout
         @Override
         public void updateUI(Location... locations)
         {
-            DecimalFormat formatter = com.forrestguice.suntimeswidget.calculator.Location.decimalDegreesFormatter();
+            DecimalFormat formatter = com.forrestguice.suntimeswidget.calculator.core.Location.decimalDegreesFormatter();
             text_locationLat.setText( formatter.format(locations[0].getLatitude()) );
             text_locationLon.setText( formatter.format(locations[0].getLongitude()) );
             text_locationAlt.setText( formatter.format(locations[0].getAltitude()) );
@@ -511,7 +511,7 @@ public class LocationConfigView extends LinearLayout
      * @param location a WidgetSettings.Location instance to update from
      */
     @SuppressLint("SetTextI18n")
-    private void updateViews(com.forrestguice.suntimeswidget.calculator.Location location)
+    private void updateViews(com.forrestguice.suntimeswidget.calculator.core.Location location)
     {
         text_locationLat.setText(location.getLatitude());
         text_locationLon.setText(location.getLongitude());
@@ -555,7 +555,7 @@ public class LocationConfigView extends LinearLayout
             spinner_locationMode.setSelection(locationMode.ordinal());
         }
 
-        com.forrestguice.suntimeswidget.calculator.Location location = WidgetSettings.loadLocationPref(context, appWidgetId);
+        com.forrestguice.suntimeswidget.calculator.core.Location location = WidgetSettings.loadLocationPref(context, appWidgetId);
         updateViews(location);
     }
 
@@ -588,12 +588,12 @@ public class LocationConfigView extends LinearLayout
         String longitude = bundle.getString(KEY_LOCATION_LONGITUDE);
         String latitude = bundle.getString(KEY_LOCATION_LATITUDE);
         String altitude = bundle.getString(KEY_LOCATION_ALTITUDE);
-        com.forrestguice.suntimeswidget.calculator.Location location;
+        com.forrestguice.suntimeswidget.calculator.core.Location location;
         if (longitude != null && latitude != null)
         {
             if (altitude != null)
-                location = new com.forrestguice.suntimeswidget.calculator.Location(label, latitude, longitude, altitude);
-            else location = new com.forrestguice.suntimeswidget.calculator.Location(label, latitude, longitude);
+                location = new com.forrestguice.suntimeswidget.calculator.core.Location(label, latitude, longitude, altitude);
+            else location = new com.forrestguice.suntimeswidget.calculator.core.Location(label, latitude, longitude);
 
         } else {
             Log.w("LocationConfigView", "Bundle contained null lat or lon; falling back to saved prefs.");
@@ -644,7 +644,7 @@ public class LocationConfigView extends LinearLayout
             String longitude = text_locationLon.getText().toString();
             String altitude = text_locationAlt.getText().toString();
             String name = text_locationName.getText().toString();
-            com.forrestguice.suntimeswidget.calculator.Location location = new com.forrestguice.suntimeswidget.calculator.Location(name, latitude, longitude, altitude, WidgetSettings.loadLengthUnitsPref(context, appWidgetId) == WidgetSettings.LengthUnit.METRIC);
+            com.forrestguice.suntimeswidget.calculator.core.Location location = new com.forrestguice.suntimeswidget.calculator.core.Location(name, latitude, longitude, altitude, WidgetSettings.loadLengthUnitsPref(context, appWidgetId) == WidgetSettings.LengthUnit.METRIC);
             WidgetSettings.saveLocationPref(context, appWidgetId, location);
             return true;
         }
@@ -899,7 +899,7 @@ public class LocationConfigView extends LinearLayout
     }
     public void copyLocationToClipboard(Context context, boolean silent)
     {
-        com.forrestguice.suntimeswidget.calculator.Location location = getLocation();
+        com.forrestguice.suntimeswidget.calculator.core.Location location = getLocation();
         String clipboardText = location.toString();
 
         if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.HONEYCOMB)
@@ -975,7 +975,7 @@ public class LocationConfigView extends LinearLayout
 
             if (cursor.getColumnCount() >= 4)
             {
-                updateViews(new com.forrestguice.suntimeswidget.calculator.Location(cursor.getString(1), cursor.getString(2), cursor.getString(3), cursor.getString(4)));
+                updateViews(new com.forrestguice.suntimeswidget.calculator.core.Location(cursor.getString(1), cursor.getString(2), cursor.getString(3), cursor.getString(4)));
             }
         }
         public void onNothingSelected(AdapterView<?> parent) {}
