+++ app/src/main/java/com/forrestguice/suntimeswidget/SuntimesConfigActivity2.java
@@ -22,7 +22,7 @@ import android.appwidget.AppWidgetManager;
 import android.content.Context;
 import android.widget.ArrayAdapter;
 
-import com.forrestguice.suntimeswidget.calculator.SuntimesCalculator;
+import com.forrestguice.suntimeswidget.calculator.core.SuntimesCalculator;
 import com.forrestguice.suntimeswidget.calculator.SuntimesCalculatorDescriptor;
 import com.forrestguice.suntimeswidget.layouts.SunPosLayout;
 import com.forrestguice.suntimeswidget.layouts.SunPosLayout_1X1_0;
