+++ app/src/main/java/com/forrestguice/suntimeswidget/calculator/SuntimesMoonData.java
@@ -20,6 +20,8 @@ package com.forrestguice.suntimeswidget.calculator;
 
 import android.content.Context;
 
+import com.forrestguice.suntimeswidget.calculator.core.SuntimesCalculator;
+
 import java.util.ArrayList;
 import java.util.Calendar;
 import java.util.HashMap;
