+++ app/src/main/java/com/forrestguice/suntimeswidget/getfix/GetFixDatabaseAdapter.java
@@ -25,7 +25,7 @@ import android.database.SQLException;
 import android.database.sqlite.SQLiteDatabase;
 import android.database.sqlite.SQLiteOpenHelper;
 
-import com.forrestguice.suntimeswidget.calculator.Location;
+import com.forrestguice.suntimeswidget.calculator.core.Location;
 
 public class GetFixDatabaseAdapter
 {
